src/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.4
asn1crypto==0.24.0
astunparse==1.6.3
beautifulsoup4==4.6.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.3
djangorestframework==3.12.4
filelock==3.0.12
firebase-admin==5.0.0
flatbuffers==1.12
future==0.15.2
gast==0.4.0
gcloud==0.17.0
GDAL==2.2.2
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
h5py==3.1.0
html5lib==0.999999999
httplib2==0.19.1
idna==2.6
importlib-metadata==4.3.0
importlib-resources==5.1.4
ipython-genutils==0.2.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
jws==0.1.3
kazam==1.4.5
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lightdm-gtk-greeter-settings==1.2.2
lxml==4.2.1
Markdown==3.3.4
MarkupSafe==1.0
matplotlib==2.1.1
msgpack==1.0.2
nbformat==4.4.0
numpy==1.19.5
oauth2client==3.0.0
oauthlib==3.1.0
openshot-qt==2.5.1.dev2
opt-einsum==3.3.0
OWSLib==0.16.0
packaging==20.9
pexpect==4.2.1
Pillow==5.1.0
plotly==2.2.3
powerline-status==2.6
proto-plus==1.18.1
protobuf==3.17.1
psutil==5.4.2
psycopg2==2.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.4.3
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pyparsing==2.4.7
pyproj==1.9.5.1
Pyrebase==3.0.27
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-jwt==2.0.1
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
requests-unixsocket==0.1.5
rsa==4.7.2
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.7
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
traitlets==4.3.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
uritemplate==3.0.1
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
virtualenv==20.4.7
webencodings==0.5
Werkzeug==2.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.4.1
